pyproject.toml
[project]
name = "aex"
version = "2.1.0"
description = "AEX v2.1 — multi-tenant AI execution governance kernel with deterministic admission, exactly-once settlement, and auditable hash-chain replay."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "Apache-2.0" }
authors = [
  { name = "Auro-rium", email = "auroriumnexus@gmail.com" }
]
keywords = ["ai", "llm", "governance", "budget", "proxy", "openai", "agents", "rate-limiting"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Typing :: Typed",
]

dependencies = [
    "fastapi>=0.110",
    "uvicorn>=0.29",
    "httpx>=0.27",
    "pydantic>=2.6",
    "psycopg[binary]>=3.2",
    "typer>=0.9",
    "rich>=13.7",
    "psutil>=5.9",
    "python-dotenv>=1.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
openai = ["openai>=1.0"]
dev = ["langgraph>=0.2"]

[project.urls]
Homepage = "https://github.com/Auro-rium/aex"
Repository = "https://github.com/Auro-rium/aex"
Issues = "https://github.com/Auro-rium/aex/issues"

[project.scripts]
aex = "aex.cli:app"

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"aex.daemon.frontend" = ["*.html"]
